feat: implement history table with loading and empty states, refactor data fetching logic
