+++ src/components/molecules/Table/index.ts
@@ -1,0 +1,2 @@
+export { TableHeaderRow } from "./TableHeaderRow";
+export { TableDataRow } from "./TableDataRow";
