+++ src/shared/utils/enums/paths.ts
@@ -12,6 +12,7 @@ type RouteParamsMap = {
   WORLD_PRAGMA: void;
   WORLD_PRAGMA_ACCOUNT: { id: string };
   PROFILE: void;
+  HISTORY: void;
 };
 
 export const ROUTE_PATHS: { [K in keyof RouteParamsMap]: Route<RouteParamsMap[K]> } = {
@@ -23,6 +24,7 @@ export const ROUTE_PATHS: { [K in keyof RouteParamsMap]: Route<RouteParamsMap[K]
   WORLD_PRAGMA_ACCOUNT: { path: "/mundo-pragma/cuenta/:id", getHref: ({ id }) => `/mundo-pragma/cuenta/${id}` },
   TRIVIA_SCRIPT: { path: "/mundo-pragma/triviascript", getHref: () => `/mundo-pragma/triviascript` },
   PROFILE: { path: "/profile", getHref: () => "/profile" },
+  HISTORY: { path: "/history", getHref: () => "/history" },
 } as const;
 
 // Routes required authentication
@@ -30,4 +32,5 @@ export const PROTECTED_ROUTES: string[] = [
   ROUTE_PATHS.HOME.getHref(),
   ROUTE_PATHS.WORLD_PRAGMA.getHref(),
   ROUTE_PATHS.PROFILE.getHref(),
+  ROUTE_PATHS.HISTORY.getHref(),
 ];
